requirements.txt
# NLP Business Intelligence — Dependencies
# Install: pip install -r requirements.txt

# ── Core ─────────────────────────────────────────────────────
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0

# ── NLP — Sentiment ──────────────────────────────────────────
vaderSentiment>=3.3.2           # fast rule-based (default backend)

# ── NLP — Transformers (optional, higher accuracy) ───────────
# Uncomment to enable DistilBERT sentiment backend
# transformers>=4.38.0
# torch>=2.0.0                  # CPU version (add +cu118 for GPU)

# ── NLP — Embeddings ─────────────────────────────────────────
sentence-transformers>=2.6.0    # all-MiniLM-L6-v2 embedding model
scikit-learn>=1.3.0             # TF-IDF fallback, LDA, PCA

# ── Dimensionality Reduction ─────────────────────────────────
umap-learn>=0.5.5               # UMAP (preferred over t-SNE)

# ── Topic Modeling ───────────────────────────────────────────
# LDA is included in scikit-learn above

# ── Optional: BERTopic backend ───────────────────────────────
# Uncomment for state-of-the-art topic modeling
# bertopic>=0.16.0
# hdbscan>=0.8.33

# ── Utilities ────────────────────────────────────────────────
python-dotenv>=1.0.0
